fix(Index): :bug: Only store data for files with metadata in them
    Closes #8
