+++ package.json
@@ -6,7 +6,8 @@
 	"scripts": {
 		"dev": "node esbuild.config.mjs",
 		"build": "node esbuild.config.mjs production",
-		"release": "standard-version"
+		"release": "standard-version",
+		"test": "jest"
 	},
 	"standard-version": {
 		"t": "",
@@ -31,15 +32,19 @@
 	"license": "MIT",
 	"devDependencies": {
 		"@types/chartjs": "^0.0.31",
+		"@types/jest": "^27.4.0",
 		"@types/json2csv": "^5.0.3",
 		"@types/lodash": "^4.14.178",
-		"@types/node": "^16.11.1",
+		"@types/luxon": "^2.0.8",
+		"@types/node": "^16.11.19",
 		"esbuild": "^0.13.12",
+		"jest": "^27.4.7",
 		"obsidian": "^0.12.17",
 		"obsidian-dataview": "^0.4.21",
 		"standard-version": "^9.3.2",
+		"ts-jest": "^27.1.2",
 		"tslib": "2.3.1",
-		"typescript": "4.4.4"
+		"typescript": "^4.5.4"
 	},
 	"dependencies": {
 		"chartjs": "^0.3.24",
