+++ src/utils.ts
@@ -5,11 +5,6 @@ import { DECIMALS } from "./const";
 import DataAnalysisPlugin from "./main";
 import { StatsModal } from "./StatsModal";
 
-export const splitAndTrim = (fields: string): string[] => {
-	if (fields === "") return [];
-	else return fields.split(",").map((str) => str.trim());
-};
-
 export function makeArr<T>(input: T | T[]): any[] {
 	return [input].flat();
 }
